refactor(config): move toml module tests to tests.rs
Relocate the ~1000-line inline `mod tests` out of mod.rs into a tests.rs
file (declared via `mod tests;`). Submodule `super` paths are unchanged, so
no test bodies needed editing.
